Pull fit method up to ModelBase
Updated imports
